refactor(main): snapshot apply / PR list / issue context model を切り出す (#224 step 3)
DIFF_APP_STATE thread_local に依存しない、UI モデル変換系のヘルパを
src/app/diff_viewer/snapshot.rs に集約。

切り出し:
- LinkedIssueDisplay (enum)
- apply_snapshot_to_ui
- build_pr_list_model
- build_issue_context_model

main.rs から不要になった import (IssueContextRecord / IssueState /
PrListState / PullRequestSummary / build_diff_file_views / excerpt /
short_sha) を整理。

差分:
- src/main.rs: 1650 → 1561 行 (-89)
- src/app/diff_viewer/snapshot.rs: 112 行 (新規)

検証: cargo clippy / cargo test (143 passed)

Refs: #224
